requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
cryptography==3.4.7
requests==2.25.1

